% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_list}
\alias{expand_list}
\title{Expand an ellipsis list.}
\usage{
expand_list(...)
}
\arguments{
\item{...}{either a list or a series of strings to be expanded}
}
\value{
a named list.
}
\description{
This function takes an ellipsis list of the form \code{...} and returns an
expanded list
}
\keyword{internal}
