% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{get_nass}
\alias{get_nass}
\title{Get data and return a data frame}
\usage{
get_nass(params, ..., as = "dataframe")
}
\arguments{
\item{params}{a named list of parameters to pass to quick stats}

\item{...}{additional parameters passed to low level functions \code{\link{get_nass.single}}
and \code{\link{get_nass.multi}}.}

\item{as}{a string indicating the desired format. Default is "dataframe".
Can also be "json" or "list".}
}
\value{
a data frame of requested data.
}
\description{
Calls nassqs_GET and nassqs_parse and returns a data frame by default.
}
\examples{
\dontrun{
params = list(COMMODITY_NAME="Corn", YEAR=2012, STATE_ALPHA="WA")
get_nass(params)
}
}

