% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{get_param_values}
\alias{get_param_values}
\title{Get all values for a specific parameter.}
\usage{
get_param_values(param, ...)
}
\arguments{
\item{param}{the name of a NASS quickstats parameter.}

\item{...}{additional parameters passed to \code{\link{nassqs_parse}}.}
}
\value{
single column data frame containing values for that parameter.
}
\description{
Equivalent to \code{nassqs_GET(list('param'=param_name))}
}

