% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{get_record_count}
\alias{get_record_count}
\title{Get a count of number of records for given parameters.}
\usage{
get_record_count(params, ...)
}
\arguments{
\item{params}{a named list of parameters and values.}

\item{...}{additional parameters passed to \code{\link{nassqs_GET}}.}
}
\value{
integer that is the number of records that fits those parameter values.
}
\description{
Returns the number of records that fit a set of parameters. Useful if your
current parameter set returns more than the 50,000 record limit.
}

