% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nass_parse.json}
\alias{nass_parse.json}
\title{Parse a request returned as JSON}
\usage{
nass_parse.json(text, ..., as)
}
\arguments{
\item{text}{the text returned from content.}

\item{...}{additional parameters passed to \code{jsonlite}.}

\item{as}{indicator of output format. One of "raw" or "list", otherwise a
data frame is returned.}
}
\value{
a data frame or a list depending on the input.
}
\description{
Parse a request returned as JSON
}

