% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs.multi}
\alias{nassqs.multi}
\title{Get data and return a data frame, can handle multiple values for parameters.}
\usage{
nassqs.multi(params, as = c("data.frame", "list", "raw"), ...)
}
\arguments{
\item{params}{a named list of parameters and the query values.}

\item{as}{determines the output. Can be "js", "list", or "dataframe".}

\item{...}{additional parameters passed to \code{\link{get_nass.single}}.}
}
\value{
a dataframe of data
}
\description{
Warning - the NASS QS API does not allow more than a single value per GET request.
To handle multiple values, this function splits the parameters into a set of single
value requests and issues a GET request for each set of values, then combines the
results into a single data frame. As a result, this can take some time to run...
}

