% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_auth}
\alias{nassqs_auth}
\title{Authenticate the user}
\usage{
nassqs_auth(key = nassqs_token())
}
\arguments{
\item{key}{api key. If not specified, looks for an env variable named NASSQS_TOKEN and if not available, asks at the console.}
}
\value{
authentication token.
}
\description{
the NASS API uses a basic html API key authorization, where the key is included in the
html call. You can add your API key in three ways:
}
\details{
(1) directly or as a variable from your R program: \code{auth = nassqs_auth(`api_key`)}
(2) by setting it in your R environment (you'll never have to enter it again).
(3) by entering it into the console when asked (it will be stored for the rest of the session.)
}
