% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{nassqs_byfips}
\alias{nassqs_byfips}
\title{Allow querying for a given set of counties based on FIPS.}
\usage{
nassqs_byfips(fips, ...)
}
\arguments{
\item{fips}{a list of 5-digit fips codes}

\item{...}{either a named list of parameters or a series of parameters to
form the query}
}
\value{
a data.frame of data for each fips code
}
\description{
This wrapper allows specifying a list of counties by FIPS code.
It iterates over each state in the list of FIPS, downloading for each
separately and then concatenating.
}
\examples{
\dontrun{
nassqs_byfips(
  fips = c("19001", "17005", "17001"),
  commodity_desc = "CORN",
  year = 2019,
  statisticcat_desc = "YIELD")
}
}
