% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_check}
\alias{nassqs_check}
\title{Check the request.}
\usage{
nassqs_check(req)
}
\arguments{
\item{req}{request result returned from quickstats}
}
\value{
parsed request result as json
}
\description{
Check the request.
}
