% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{nassqs_fields}
\alias{nassqs_fields}
\title{Deprecated: Return list of NASS QS parameters.}
\usage{
nassqs_fields(...)
}
\arguments{
\item{...}{a parameter, series of parameters, or a list of parameters that
you would like a description of. If missing, a list of all available
parameters is returned.}
}
\description{
Deprecated. Use \code{\link[=nassqs_params]{nassqs_params()}} instead.
}
