% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_GET}
\alias{nassqs_GET}
\title{Issue a GET request to the NASS API}
\usage{
nassqs_GET(params, ..., path = "api_GET",
  base_url = "http://quickstats.nass.usda.gov/api/", key = nassqs_auth(),
  format = "JSON")
}
\arguments{
\item{params}{a named list of values to be queried.}

\item{path}{the api path. Can be "api_GET", "get_param_values", or "get_counts"}

\item{format}{format of returned data. JSON by default, but can also be XML or CSV.}
}
\description{
Issue a GET request to the NASS API
}

