% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nassqs_param_values}
\alias{nassqs_param_values}
\title{Depreciated: Get all values for a specific parameter.}
\usage{
nassqs_param_values(param, ...)
}
\description{
Use \code{nassqs_field_values()} instead.
}
