% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_parse.json}
\alias{nassqs_parse.json}
\title{Parse a request returned as JSON.}
\usage{
nassqs_parse.json(text, as = c("data.frame", "list", "raw"), ...)
}
\arguments{
\item{text}{the text returned from \code{content()}.}

\item{as}{indicator of output format..}

\item{...}{additional parameters passed to \code{fromJSON()}.}
}
\value{
a data frame or a list depending on the input.
}
\description{
Parse a request returned as JSON.
}

