% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_parse.xml}
\alias{nassqs_parse.xml}
\title{Parse a NASS QS request returned as XML.}
\usage{
nassqs_parse.xml(text, as = c("data.frame", "list", "raw"), ...)
}
\arguments{
\item{text}{the text returned from \code{content()}.}

\item{as}{indicator of output format.}

\item{...}{additional parameters passed to \code{fromJSON()}.}
}
\value{
a data frame or a list depending on the the "as" parameter.
}
\description{
Parse a NASS QS request returned as XML.
}

