% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_token}
\alias{nassqs_token}
\title{Get the user's API key from the environment}
\usage{
nassqs_token(force = FALSE)
}
\arguments{
\item{force}{a boolean to force asking in the console.}
}
\value{
the api key.
}
\description{
First checks if the NASSQS_TOKEN environmental variable is set, and if so,
returns it. Otherwise asks the console. If not an interactive session,
fails with error msg.
}

