% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{nassqs_yield}
\alias{nassqs_yield}
\title{Wrapper to get NASS yields for a given crop}
\usage{
nassqs_yield(params, ...)
}
\arguments{
\item{params}{a named list of parameters}

\item{...}{additional parameters passed to \code{\link{get_nass}}}
}
\value{
a dataset of NASSQS data.
}
\description{
Takes a list of parameters and adds a parameter to specify yield.
This is just a convenience function so that you don't have to add
a yield parameter every time you want to fetch yields.
}

