% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{nassqs_yields}
\alias{nassqs_yields}
\title{Get yield records for a specified crop.}
\usage{
nassqs_yields(...)
}
\arguments{
\item{...}{either a named list of parameters or a series of parameters to
form the query}
}
\value{
a data.frame of yields data
}
\description{
Returns yields for other specified parameters. This function is intended to
simplify common requests.
}
\examples{
\dontrun{
  # Get yields for wheat in 2012, all geographies
  params <- list(
    commodity_desc = "WHEAT", 
    year = "2012", 
    agg_level_desc = "STATE",
    state_alpha = "WA")
    
  yields <- nassqs_yields(params)
  head(yields)
}
}
