% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qs_fields}
\alias{qs_fields}
\title{NASS QUICKSTATS field names}
\format{\preformatted{List of 39
 $ : chr "agg_level_desc"
 $ : chr "asd_code"
 $ : chr "asd_desc"
 $ : chr "begin_code"
 $ : chr "class_desc"
 $ : chr "commodity_desc"
 $ : chr "congr_district_code"
 $ : chr "country_code"
 $ : chr "country_name"
 $ : chr "county_ansi"
 $ : chr "county_code"
 $ : chr "county_name"
 $ : chr "CV"
 $ : chr "domaincat_desc"
 $ : chr "domain_desc"
 $ : chr "end_code"
 $ : chr "freq_desc"
 $ : chr "group_desc"
 $ : chr "load_time"
 $ : chr "location_desc"
 $ : chr "prodn_practice_desc"
 $ : chr "reference_period_desc"
 $ : chr "region_desc"
 $ : chr "sector_desc"
 $ : chr "short_desc"
 $ : chr "state_alpha"
 $ : chr "state_ansi"
 $ : chr "state_name"
 $ : chr "state_fips_code"
 $ : chr "statisticcat_desc"
 $ : chr "source_desc"
 $ : chr "unit_desc"
 $ : chr "util_practice_desc"
 $ : chr "Value"
 $ : chr "watershed_code"
 $ : chr "watershed_desc"
 $ : chr "week_ending"
 $ : chr "year"
 $ : chr "zip_5"
}}
\usage{
qs_fields
}
\description{
NASS QUICKSTATS field names
}
\keyword{datasets}

