% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_list}
\alias{expand_list}
\title{Expand an ellipsis list.}
\usage{
expand_list(listx)
}
\arguments{
\item{listx}{a list to be expanded.}
}
\value{
an expanded list.
}
\description{
Expand an ellipsis list.
}
