% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\docType{package}
\name{nassqs}
\alias{nassqs}
\alias{nassqs-package}
\title{nassqs-package: Use the NASS Quickstats API from R}
\description{
This package wraps around the NASS Quickstats API so you can make R calls
to fetch data. Based on the httr api package guide.
}
\examples{
\dontrun{get.user.repositories("potterzot")}
}
\author{
Nicholas Potter
}
\seealso{
\code{jsonlite}

\url{http://quickstats.nass.usda.gov/api}
}
\keyword{nassqs-package}
\keyword{package}

