% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs}
\alias{nassqs}
\title{Get data and return a data frame}
\usage{
nassqs(params, ...)
}
\arguments{
\item{params}{a named list of parameters to pass to quick stats}

\item{...}{additional parameters passed to the low level function \code{\link{nassqs.single}}.}
}
\value{
a data frame of requested data.
}
\description{
Calls nassqs_GET and nassqs_parse and returns a data frame by default.
}
\examples{
\dontrun{
params = list(COMMODITY_NAME="CORN", YEAR=2012, STATE_ALPHA="WA")
nassqs(params)
}
}
