% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs}
\alias{nassqs}
\title{Get data and return a data frame}
\usage{
nassqs(params, ...)
}
\arguments{
\item{params}{a named list of parameters to pass to quick stats}

\item{...}{additional parameters passed to low level functions \code{\link{get_nass.single}}
and \code{\link{get_nass.multi}}.}
}
\value{
a data frame of requested data.
}
\description{
Calls nassqs_GET and nassqs_parse and returns a data frame by default.
}
\examples{
\dontrun{
params = list(COMMODITY_NAME="Corn", YEAR=2012, STATE_ALPHA="WA")
get_nass(params)
}
}

