% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs.single}
\alias{nassqs.single}
\title{Get data and return a data frame, in which each parameter selected has only one value.}
\usage{
nassqs.single(params, ..., as = "dataframe")
}
\arguments{
\item{params}{a named list of parameters to pass to quick stats. One value per parameter.}

\item{...}{additional parameters passed to \code{\link{nassqs_parse}}.}

\item{as}{a string naming the format the data should be returned as. Default is dataframe.}
}
\value{
a data frame of requested data.
}
\description{
Calls nassqs_GET and nassqs_parse to return a data frame
}

