% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{nassqs_check}
\alias{nassqs_check}
\title{Check the response.}
\usage{
nassqs_check(response)
}
\arguments{
\item{response}{a \code{\link[httr:GET]{httr::GET()}} request result returned from the API.}
}
\value{
nothing if check is passed, or an informative error if not passed.
}
\description{
Check that the response is valid, i.e. that it doesn't exceed 50,000 records
and that all the parameter values are valid. This is used to ensure that
the query is valid before querying to reduce wait times before receiving an
error.
}
