% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_GET}
\alias{nassqs_GET}
\title{Issue a GET request to the NASS API
http://quickstats.nass.usda.gov/api
This is the core function, which several other rnassqs functions use to request data.}
\usage{
nassqs_GET(params, api_path = c("api_GET", "get_param_values",
  "get_counts"), base_url = "https://quickstats.nass.usda.gov/api/",
  key = nassqs_auth(), debug = FALSE, format = c("JSON", "XML",
  "CSV"))
}
\arguments{
\item{params}{a named list of values to be queried.}

\item{api_path}{the api path. Can be "api_GET", "get_param_values", or 
"get_counts".}

\item{base_url}{the base api url. This should probably never be changed.}

\item{key}{your api key. If not provided the function will check for an env 
var and if not found, will prompt for your api key.}

\item{debug}{(logical) if TRUE, returns the URL and makes no API call.}

\item{format}{format of returned data. JSON by default, but can also be XML 
or CSV. Can also be set as a parameter.}
}
\value{
data returned in the format specified.
}
\description{
Issue a GET request to the NASS API
http://quickstats.nass.usda.gov/api
This is the core function, which several other rnassqs functions use to request data.
}
