% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nassqs.R
\name{nassqs_parse}
\alias{nassqs_parse}
\title{Parse the returned request}
\usage{
nassqs_parse(req)
}
\arguments{
\item{req}{the GET request}
}
\value{
an object formatted as desired
}
\description{
Parse the returned request
}

