% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nassqs_yield}
\alias{nassqs_yield}
\title{Get yield records for a specified crop.}
\usage{
nassqs_yield(params, ...)
}
\arguments{
\item{params}{a named list of parameters}

\item{...}{additional parameters passed to \code{\link{nassqs}}}
}
\value{
a dataset of NASSQS data.
}
\description{
Convenience function to take a list of parameters and add a parameter to
specify yield. Equivalent to \code{nassqs(list(..., "statisticcat_desc" = "YIELD"))}
}
\examples{
\donttest{
  #get yields for wheat in 2012, all geographies
  params = list(commodity_desc="WHEAT", year="2012")
  nassqs_yield(params)
}
}
