% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nassqs.R
\docType{package}
\name{rnassqs-package}
\alias{rnassqs-package}
\alias{rnassqs}
\title{rnassqs-package: Use the NASS Quickstats API from R.}
\description{
rnassqs is a wrapper for the NASS Quickstats API to enable getting NASS
QuickStats data from R directly.  Based on the httr api package guide.
}
\details{
The functions in this package facilitate getting data from NASS QuickStats.
It handles the API key checking and storage, authorization, and fetching of data.
}
\examples{
\dontrun{

}
}
\references{
\url{http://quickstats.nass.usda.gov}
}
\seealso{
\url{http://quickstats.nass.usda.gov/api}
}
\author{
Nicholas Potter
}
\keyword{package}
\keyword{rnassqs-package}
